% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_clean_simpleaf_barcodes.R
\name{clean_simpleaf_barcodes}
\alias{clean_simpleaf_barcodes}
\title{Clean simpleaf Cell Names and Export Barcode Mapping}
\usage{
clean_simpleaf_barcodes(
  seurat_obj,
  output_dir = NULL,
  file_name = "b2sample.tsv"
)
}
\arguments{
\item{seurat_obj}{A \code{Seurat} object.}

\item{output_dir}{Character. Optional directory to save the \code{b2sample.tsv} file.}

\item{file_name}{Character. Name of the output TSV mapping file. Default is \code{"b2sample.tsv"}.}
}
\value{
A renamed \code{Seurat} object.
}
\description{
Standardizes cell names in a Seurat object generated from simpleaf in nf-core/scrnaseq by
removing the sample suffix (e.g., keeping only the barcode sequence).
Extracts the sample name and exports a 2-column TSV (Barcode, Sample).
}
